refactor: Segregrate benches into preparation and execution parts (#1650)
This PR introduces `Bench` trait, which refactors the way in which we
bench. It is intended to customise benches, having a preparation phase,
which might be needlessly expensive, and one which not want to
benchmark. With `Bench` trait, we can implement the preparation phase
with `prepare` method, and include the actual benchmark in `execute`.
This simplifies the PR: https://github.com/0xmozak/mozak-vm/pull/1651

---------

Co-authored-by: Matthias Goergens <matthias.goergens@gmail.com>