refactor+fix: redo all the recproof tests (#1679)
This is a massive refactor of all the recproof test code. The basic
premise is to create a single set of underlying events and programs that
all tests use. The outputs of each test are then cached and then used in
any future stages that need them.

I have also removed all parameters from `prove` functions which only
existed for testing purposes, keeping them only in `prove_unsafe`
functions for negative testing and opting instead to check the
`public_input` values of proofs in the tests (much cleaner).

One nice bonus is this actually makes the tests faster as we don't have
to duplicate proof creation tests require for setup. Unfortunately I
have immediately ruined that by increasing the number of tests from ~100
to ~300, causing the time to go from ~45s to ~55s.