refactor: remove all hashing from state update driver (#1704)
An earlier refactor removed the need to pass implicit values into the
proofs. This cleans up all the unused node code resulting from that
refactor.

Also it turns on parallelism for tests and add a few comments to get rid
of some clippy warnings.

There is no change to functionality.