feat: get `self_prog_id` by hashing merkle caps of `ProgramRom`  and `ElfMemoryInit`  tables and `entry_point` (#1607)
For us, an ELF and the initial state of its corresponding program can be
uniquely identified by its `ro_code`, `ro_memory`, `rw_memory` and
`entry_point` data from the ELF. This data is tied to the VM(starky)
proof for an execution of ELF as follows.
- `ro_code` is stored in table `ProgramRom`
- `ro_memory`, `rw_memory` lies in `ElfMemoryInit`
- `entry_point` is same as `pc` of the first row of `Cpu` table, and is
public input to its proof.

So we can derive a `program_hash` for the ELF that is tied to every
possible execution proof, by hashing the merkle caps of `ProgramRom` and
`ElfMemoryInit`, and `entry_point`. Note that this can be computed
independently of the proof by an entity which has access only to the ELF
of the program (or code), and hence can be treated as identifier
`self_prog_id` for the ELF. Further, the prover who wants to prove an
execution of such ELF, can compute `program_hash` inside the circuit,
and make it public through the final recursive circuit. Hence an outside
entity, who has successfully verified the recursive proof with
`self_prog_id` as one of the public inputs, can be assured that the
proof indeed corresponds to the ELF.

The PR provides code to compute `self_prog_id` from both the perspective
of recursive prover, as well as an entity who has access to ELF alone.
That is,
- The recursive prover has an additional public input
`program_hash_as_bytes` which is tied to the hash of merkle caps and
entry point, through recursive circuit.
- Through Cli, we can compute `self_prog_id` for a given ELF, and can
verify a recursive proof using it as one of the pubilc inputs.

The code to test the above interaction has been included in
`integration_test.rs`. It should also simplify transaction bundling, as
we no longer need to keep a map between ELF path and its
`ProgramIdentifier`

---------

Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>