feat(wallet): implement signature approval (#1709)
We approve signatures by asserting the following equality:

```rust
hash(private_key) == public_key
```

Where `hash` can be any hard to invert function (in this case it's the
poseidon2 hash).

For the purpose of this PR, we randomly generate a private key, which we
then hash to obtain a public key. We write the private key to our
private tape during native execution, which we then read during guest
execution. A poseidon2 ecall helps us prove that we know the pre-image.

We may use our `signatures` crate in future as an optimization, once we
link it to our SDK.