chore: replace all usages of `access_unchecked` with `access`
This requires us to enable the `bytecheck` features of `rkyv` and makes
use of the `CheckBytes` trait to do byte validation when dealing with
`rkyv` types.

In future, if we wish to bring back the unsafe API for those who want performance,
we could feature gate that in our sdk.