Specialise Poseidon2 constants to Goldilocks field (#1784)
Previously part of our constraints for Poseidon2 tried to be generic in
the field. But actually using a different field within plonky2 would
require a lot of infrastructure changes anyway. So we are hard-coding
Goldilocks here, to make the types simpler.

This also simplifies the next PR that restores constraint debugging.

Extracted from https://github.com/0xmozak/mozak-vm/pull/1756