feat: Use build script to dump self-prog-id + simplify cli commands  (#1763)
main features of the PR:
- Added crate `build-scripts` to hold all the scripts used for building.
The main script introduced in PR is `dump_self_prog_id(example: &str)`,
which is used to compute `self_prog_id` from ELF, and store it as a `pub
const &str` with name `{example}_SELF_PROG_ID`. A separate binary
`dump-self-prog-id` is introduced in `cli` so that we don't have to
build entire `mozak-cli` (which takes 2x amount of time).
- Each (fully featured for now) example contains an additional crate
`elf_data`, whose build actually triggers the above build script. It is
also useful to access program ids of other examples, like a `token`
example can access program id of `wallet` through
`wallet_elf_data::WALLET_SELF_PROG_ID`.
- Now that `self_prog_id` are no longer hardcoded, we don't need to pass
them through arguments for cli commands. Corresponding commands are
updated to reflect this.
- Now that above is taken care of, and since system tape path is fixed
relative to each example, we are able to simplify the `mozakvm-run`
command to work without any arguments.
- `VerifyRecursiveProof` command takes in additional input `program-id`
that allows it to check that the proof indeed corresponds to the program
verifer meant to verify proof of.

---------

Co-authored-by: bing <bingcicle@mozak.com>
Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>